+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+## [2.0] (2021-05-01)
+
 ### ReactiveMap refactored
 
 `ReactiveMap` renamed to `ReactiveMutableMap`.
@@ -153,6 +155,7 @@ You should use `mutableMap` and `mutableList` for mutable versions of collection
 
 Public release
 
-[unreleased]: https://github.com/RedMadRobot/mapmemory/compare/v2.0-rc1...main
+[unreleased]: https://github.com/RedMadRobot/mapmemory/compare/v2.0...main
+[2.0-rc1]: https://github.com/RedMadRobot/mapmemory/compare/v2.0-rc1..v2.0
 [2.0-rc1]: https://github.com/RedMadRobot/mapmemory/compare/v1.1...v2.0-rc1
 [1.1]: https://github.com/RedMadRobot/mapmemory/compare/v1.0...v1.1